requirements.txt
pywebview>=5.0
yt-dlp>=2025.1.0
gallery-dl>=1.31.0
browser-cookie3>=0.20
requests>=2.31
